Reuse Assisntants settings component (#678)
* Update Assisntants settings

* format
